Changed CheckIntersection login, removed all memory leaks, Compiles with -Wall -Wextra -Werror, updated makefile

+++ Game/Game.h
@@ -118,7 +118,7 @@ private:
 	 *	Speed of game
 	 */
 
-	size_t game_speed;
+	long long game_speed;
 
 	/*
 	 *	Time period to change fruit position
